build(tests): enable defensive compiler warnings
Enable extra warnings for tests only to catch issues early
without enforcing them on library consumers:

GCC/Clang: -Wshadow, -Wconversion, -Wsign-conversion,
-Wnon-virtual-dtor, -Wold-style-cast, -Wcast-align,
-Wnull-dereference, -Wdouble-promotion, -Wformat=2,
-Wimplicit-fallthrough, plus GCC-specific flags.

MSVC: additional /wNNNNN warnings enabled.

Fix all warnings surfaced by the new flags across test files
and one library header (int vs size_t in emitLog).

+++ tests/test_protocol.cpp
@@ -70,7 +70,7 @@ TEST_CASE("protocol")
         REQUIRE(msg[2] == args);
     }
     SECTION("error") {
-        json msg = Protocol::errorMessage(msgType, requestId, error);
+        json msg = Protocol::errorMessage(msgType, static_cast<int>(requestId), error);
         REQUIRE(msg[0] == MsgType::Error);
         REQUIRE(msg[1] == msgType);
         REQUIRE(msg[2] == requestId);
